+++ app/src/main/java/org/oppia/android/app/hintsandsolution/HintsAndSolutionDialogFragmentPresenter.kt
@@ -7,12 +7,18 @@ import androidx.fragment.app.Fragment
 import org.oppia.android.R
 import org.oppia.android.app.fragment.FragmentScope
 import org.oppia.android.app.model.State
+import org.oppia.android.app.recyclerview.BindableAdapter
 import org.oppia.android.app.viewmodel.ViewModelProvider
 import org.oppia.android.databinding.HintsAndSolutionFragmentBinding
+import org.oppia.android.databinding.HintsDividerBinding
+import org.oppia.android.databinding.HintsSummaryBinding
+import org.oppia.android.databinding.SolutionSummaryBinding
 import org.oppia.android.util.gcsresource.DefaultResourceBucketName
 import org.oppia.android.util.parser.html.ExplorationHtmlParserEntityType
 import org.oppia.android.util.parser.html.HtmlParser
 import javax.inject.Inject
+
+const val TAG_REVEAL_SOLUTION_DIALOG = "REVEAL_SOLUTION_DIALOG"
 
 /** Presenter for [HintsAndSolutionDialogFragment], sets up bindings from ViewModel. */
 @FragmentScope
@@ -30,9 +36,10 @@ class HintsAndSolutionDialogFragmentPresenter @Inject constructor(
   private var solutionIndex: Int? = null
   private var isSolutionRevealed: Boolean? = null
   private lateinit var expandedHintListIndexListener: ExpandedHintListIndexListener
-  private lateinit var hintsAndSolutionAdapter: HintsAndSolutionAdapter
   private lateinit var binding: HintsAndSolutionFragmentBinding
   private lateinit var state: State
+  private lateinit var itemList: List<HintsAndSolutionItemViewModel>
+  private lateinit var bindingAdapter: BindableAdapter<HintsAndSolutionItemViewModel>
 
   val viewModel by lazy {
     getHintsAndSolutionViewModel()
@@ -93,24 +100,11 @@ class HintsAndSolutionDialogFragmentPresenter @Inject constructor(
       viewModel.setHintsList(state.interaction.hintList)
       viewModel.setSolution(state.interaction.solution)
 
-      hintsAndSolutionAdapter =
-        HintsAndSolutionAdapter(
-          fragment,
-          viewModel.processHintList(),
-          expandedHintListIndexListener,
-          currentExpandedHintListIndex,
-          viewModel.explorationId.get(),
-          htmlParserFactory,
-          resourceBucketName,
-          entityType,
-          index,
-          isHintRevealed,
-          solutionIndex,
-          isSolutionRevealed
-        )
+      itemList = viewModel.processHintList()
 
       binding.hintsAndSolutionRecyclerView.apply {
-        adapter = hintsAndSolutionAdapter
+        bindingAdapter = createRecyclerViewAdapter()
+        adapter = bindingAdapter
       }
       if (viewModel.newAvailableHintIndex.get() != -1) {
         handleNewAvailableHint(viewModel.newAvailableHintIndex.get())
@@ -121,8 +115,216 @@ class HintsAndSolutionDialogFragmentPresenter @Inject constructor(
     }
   }
 
+  private enum class ViewType {
+    VIEW_TYPE_HINT_ITEM,
+    VIEW_TYPE_SOLUTION_ITEM,
+    VIEW_TYPE_HINTS_DIVIDER_ITEM
+  }
+
+  private fun createRecyclerViewAdapter(): BindableAdapter<HintsAndSolutionItemViewModel> {
+    return BindableAdapter.MultiTypeBuilder
+      .newBuilder<HintsAndSolutionItemViewModel, ViewType> { viewModel ->
+        when (viewModel) {
+          is HintsViewModel -> ViewType.VIEW_TYPE_HINT_ITEM
+          is SolutionViewModel -> ViewType.VIEW_TYPE_SOLUTION_ITEM
+          is HintsDividerViewModel -> ViewType.VIEW_TYPE_HINTS_DIVIDER_ITEM
+          else -> throw IllegalArgumentException("Encountered unexpected view model: $viewModel")
+        }
+      }
+      .registerViewDataBinder(
+        viewType = ViewType.VIEW_TYPE_HINT_ITEM,
+        inflateDataBinding = HintsSummaryBinding::inflate,
+        setViewModel = this::bindHintsViewModel,
+        transformViewModel = { it as HintsViewModel }
+      )
+      .registerViewDataBinder(
+        viewType = ViewType.VIEW_TYPE_SOLUTION_ITEM,
+        inflateDataBinding = SolutionSummaryBinding::inflate,
+        setViewModel = this::bindSolutionViewModel,
+        transformViewModel = { it as SolutionViewModel }
+      )
+      .registerViewBinder(
+        viewType = ViewType.VIEW_TYPE_HINTS_DIVIDER_ITEM,
+        inflateView = { parent ->
+          HintsDividerBinding.inflate(
+            LayoutInflater.from(parent.context),
+            parent,
+            /* attachToParent= */ false
+          ).root
+        },
+        bindView = { view, viewModel -> }
+      )
+      .build()
+  }
+
+  private fun bindHintsViewModel(
+    binding: HintsSummaryBinding,
+    hintsViewModel: HintsViewModel
+  ) {
+
+    binding.viewModel = hintsViewModel
+
+    val position: Int = itemList.indexOf(hintsViewModel)
+
+    var isHintListVisible = false
+    if (currentExpandedHintListIndex != null) {
+      isHintListVisible = currentExpandedHintListIndex!! == position
+    }
+    binding.isListExpanded = isHintListVisible
+
+    index?.let {
+      isHintRevealed?.let {
+        if (index == position && isHintRevealed!!) {
+          hintsViewModel.isHintRevealed.set(true)
+        }
+      }
+    }
+
+    binding.hintTitle.text = hintsViewModel.title.get()!!.replace("_", " ").capitalize()
+    binding.hintsAndSolutionSummary.text =
+      htmlParserFactory.create(
+        resourceBucketName, entityType, hintsViewModel.explorationId.get()!!, /* imageCenterAlign= */ true
+      ).parseOppiaHtml(
+        hintsViewModel.hintsAndSolutionSummary.get()!!, binding.hintsAndSolutionSummary
+      )
+
+    if (hintsViewModel.hintCanBeRevealed.get()!!) {
+      binding.root.visibility = View.VISIBLE
+      binding.revealHintButton.setOnClickListener {
+        hintsViewModel.isHintRevealed.set(true)
+        expandedHintListIndexListener.onRevealHintClicked(position, /* isHintRevealed= */ true)
+        (fragment.requireActivity() as? RevealHintListener)?.revealHint(
+          saveUserChoice = true,
+          hintIndex = position / 2
+        )
+        val previousIndex: Int? = currentExpandedHintListIndex
+        currentExpandedHintListIndex =
+          if (currentExpandedHintListIndex != null && currentExpandedHintListIndex == position) {
+            null
+          } else {
+            position
+          }
+        expandedHintListIndexListener.onExpandListIconClicked(currentExpandedHintListIndex)
+        if (previousIndex != null && previousIndex != currentExpandedHintListIndex) {
+          bindingAdapter.notifyItemChanged(previousIndex)
+        }
+      }
+    }
+
+    binding.root.setOnClickListener {
+      if (hintsViewModel.isHintRevealed.get()!!) {
+        val previousIndex: Int? = currentExpandedHintListIndex
+        currentExpandedHintListIndex =
+          if (currentExpandedHintListIndex != null && currentExpandedHintListIndex == position) {
+            null
+          } else {
+            position
+          }
+        expandedHintListIndexListener.onExpandListIconClicked(currentExpandedHintListIndex)
+        if (previousIndex != null &&
+          currentExpandedHintListIndex != null &&
+          previousIndex == currentExpandedHintListIndex
+        ) {
+          bindingAdapter.notifyItemChanged(currentExpandedHintListIndex!!)
+        } else {
+          if (previousIndex != null) {
+            bindingAdapter.notifyItemChanged(previousIndex)
+          }
+          if (currentExpandedHintListIndex != null) {
+            bindingAdapter.notifyItemChanged(currentExpandedHintListIndex!!)
+          }
+        }
+      }
+    }
+  }
+
+  private fun bindSolutionViewModel(
+    binding: SolutionSummaryBinding,
+    solutionViewModel: SolutionViewModel
+  ) {
+
+    binding.viewModel = solutionViewModel
+
+    val position: Int = itemList.indexOf(solutionViewModel)
+
+    var isHintListVisible = false
+    if (currentExpandedHintListIndex != null) {
+      isHintListVisible = currentExpandedHintListIndex!! == position
+    }
+    binding.isListExpanded = isHintListVisible
+
+    solutionIndex?.let {
+      isSolutionRevealed?.let {
+        if (solutionIndex == position && isSolutionRevealed!!) {
+          solutionViewModel.isSolutionRevealed.set(true)
+        }
+      }
+    }
+
+    binding.solutionTitle.text = solutionViewModel.title.get()!!.capitalize()
+    // TODO(#1050): Update to display answers for any answer type.
+    if (solutionViewModel.correctAnswer.get().isNullOrEmpty()) {
+      binding.solutionCorrectAnswer.text =
+        """${solutionViewModel.numerator.get()}/${solutionViewModel.denominator.get()}"""
+    } else {
+      binding.solutionCorrectAnswer.text = solutionViewModel.correctAnswer.get()
+    }
+    binding.solutionSummary.text = htmlParserFactory.create(
+      resourceBucketName, entityType, viewModel.explorationId.get()!!, /* imageCenterAlign= */ true
+    ).parseOppiaHtml(
+      solutionViewModel.solutionSummary.get()!!, binding.solutionSummary
+    )
+
+    if (solutionViewModel.solutionCanBeRevealed.get()!!) {
+      binding.root.visibility = View.VISIBLE
+      binding.revealSolutionButton.setOnClickListener {
+        showRevealSolutionDialogFragment()
+      }
+    }
+
+    binding.root.setOnClickListener {
+      if (solutionViewModel.isSolutionRevealed.get()!!) {
+        val previousIndex: Int? = currentExpandedHintListIndex
+        currentExpandedHintListIndex =
+          if (currentExpandedHintListIndex != null && currentExpandedHintListIndex == position) {
+            null
+          } else {
+            position
+          }
+        expandedHintListIndexListener.onExpandListIconClicked(currentExpandedHintListIndex)
+        if (previousIndex != null &&
+          currentExpandedHintListIndex != null &&
+          previousIndex == currentExpandedHintListIndex
+        ) {
+          bindingAdapter.notifyItemChanged(currentExpandedHintListIndex!!)
+        } else {
+          if (previousIndex != null) {
+            bindingAdapter.notifyItemChanged(previousIndex)
+          }
+          if (currentExpandedHintListIndex != null) {
+            bindingAdapter.notifyItemChanged(currentExpandedHintListIndex!!)
+          }
+        }
+      }
+    }
+  }
+
   private fun handleAllHintsExhausted(allHintsExhausted: Boolean) {
-    hintsAndSolutionAdapter.setSolutionCanBeRevealed(allHintsExhausted)
+    if (itemList[itemList.size - 2] is SolutionViewModel) {
+      val solutionViewModel = itemList[itemList.size - 2] as SolutionViewModel
+      solutionViewModel.solutionCanBeRevealed.set(allHintsExhausted)
+      bindingAdapter.notifyItemChanged(itemList.size - 2)
+    }
+  }
+
+  private fun showRevealSolutionDialogFragment() {
+    val previousFragment =
+      fragment.childFragmentManager.findFragmentByTag(TAG_REVEAL_SOLUTION_DIALOG)
+    if (previousFragment != null) {
+      fragment.childFragmentManager.beginTransaction().remove(previousFragment).commitNow()
+    }
+    val dialogFragment = RevealSolutionDialogFragment.newInstance()
+    dialogFragment.showNow(fragment.childFragmentManager, TAG_REVEAL_SOLUTION_DIALOG)
   }
 
   private fun getHintsAndSolutionViewModel(): HintsViewModel {
@@ -130,17 +332,42 @@ class HintsAndSolutionDialogFragmentPresenter @Inject constructor(
   }
 
   fun handleRevealSolution() {
-    hintsAndSolutionAdapter.setRevealSolution()
+    if (itemList[itemList.size - 2] is SolutionViewModel) {
+      val solutionViewModel = itemList[itemList.size - 2] as SolutionViewModel
+      solutionViewModel.isSolutionRevealed.set(true)
+      expandedHintListIndexListener.onRevealSolutionClicked(
+        /* solutionIndex= */ itemList.size - 2,
+        /* isSolutionRevealed= */ true
+      )
+      (fragment.requireActivity() as? RevealSolutionInterface)?.revealSolution()
+      val previousIndex: Int? = currentExpandedHintListIndex
+      currentExpandedHintListIndex =
+        if (currentExpandedHintListIndex != null &&
+          currentExpandedHintListIndex == itemList.size - 2
+        ) {
+          null
+        } else {
+          itemList.size - 2
+        }
+      expandedHintListIndexListener.onExpandListIconClicked(currentExpandedHintListIndex)
+      if (previousIndex != null && previousIndex != currentExpandedHintListIndex) {
+        bindingAdapter.notifyItemChanged(previousIndex)
+      }
+    }
   }
 
   private fun handleNewAvailableHint(hintIndex: Int?) {
-    hintsAndSolutionAdapter.setNewHintIsAvailable(hintIndex!!)
+    if (itemList[hintIndex!!] is HintsViewModel) {
+      val hintsViewModel = itemList[hintIndex!!] as HintsViewModel
+      hintsViewModel.hintCanBeRevealed.set(true)
+      bindingAdapter.notifyItemChanged(hintIndex)
+    }
   }
 
   fun onExpandClicked(index: Int?) {
     currentExpandedHintListIndex = index
     if (index != null)
-      hintsAndSolutionAdapter.notifyItemChanged(index)
+      bindingAdapter.notifyItemChanged(index)
   }
 
   fun onRevealHintClicked(index: Int?, isHintRevealed: Boolean?) {
