+++ app/src/main/res/layout-land/hints_and_solution_fragment.xml
@@ -47,6 +47,7 @@
         android:paddingTop="24dp"
         android:paddingBottom="120dp"
         android:scrollbars="none"
+        app:list="@{viewModel.itemList}"
         app:layoutManager="androidx.recyclerview.widget.LinearLayoutManager"
         app:layout_constraintBottom_toBottomOf="parent"
         app:layout_constraintEnd_toEndOf="parent"
